Go all in on `isolatedModules`
